fix: :bug: only use `DOMPurify.sanitize` when `DOMPurify` is supported
DOMPurify returns early in its core business logic instantiation so when
a DOM is not available ostensibly because it relies on the DOM to parse
the given HTML string. To support React Native, we need to only sanitize
the HTML string when `DOMPurify` supports it.

This allows `@bonfhir/react`'s `FhirValue` component to be used in React
Native applications though by making the application developer
responsible for making sure the rendered content is safe in such a case.
Rendering Markdown content as HTML didn't work at all before in React
Native so this is a positive change though I haven't fully explored the
real vulnerability of rendering Markdown or HTML into a React Native
app.